pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Texture processing pipeline for Penumbra"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0",
    "pillow>=11.0",
    "tqdm>=4.66",
    "httpx>=0.28",
    "tomli-w>=1.1",
]

[project.scripts]
pipeline = "pipeline.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "jupyterlab>=4.0",
    "ipykernel>=6.29",
    "matplotlib>=3.9",
]
